MarkdownBear: Move to markdown module
MarkdownBear is an own language specification, though you write
texts in it, this doesn't really belong to natural language. Thus
the move from bears.natural_language.MarkdownBear to
bears.markdown.MarkdownBear.

Fixes https://github.com/coala-analyzer/coala-bears/issues/208
